+++ lib/boost.php
@@ -1510,14 +1510,18 @@ function boost_rrdtool_function_create($local_data_id, $show_source, &$rrdtool_p
     } else {
         $success = rrdtool_execute("create $quoted_path $create_ds$create_rra", false, RRDTOOL_OUTPUT_STDOUT, $rrdtool_pipe, 'BOOST');
 
-        if ($config['cacti_server_os'] != 'win32' && posix_getuid() == 0
-            && rrdtool_ownership_allowed($data_source_path, $config['rra_path'], 'BOOST')) {
-            if (!lchown($data_source_path, (int) $owner_id)) {
+        if ($config['cacti_server_os'] != 'win32' && posix_getuid() == 0) {
+            if (!file_exists($data_source_path)) {
                 cacti_log("WARNING: Unable to set owner for '" . $data_source_path . "'", false, 'BOOST');
-            }
-
-            if (!lchgrp($data_source_path, (int) $group_id)) {
                 cacti_log("WARNING: Unable to set group for '" . $data_source_path . "'", false, 'BOOST');
+            } elseif (($real_path = rrdtool_ownership_path($data_source_path, $config['rra_path'], 'BOOST')) !== false) {
+                if (!lchown($real_path, (int) $owner_id)) {
+                    cacti_log("WARNING: Unable to set owner for '" . $data_source_path . "'", false, 'BOOST');
+                }
+
+                if (!lchgrp($real_path, (int) $group_id)) {
+                    cacti_log("WARNING: Unable to set group for '" . $data_source_path . "'", false, 'BOOST');
+                }
             }
         }